Update make_a_db.py
this is a file which allows users to create default layout databases. There's A LOT of room for improvement, the code is insecure, because of the amount of string formatting, it's not very user-friendly, and can only run on terminal. All files created must have a .db extension otherwise it can't be read. This was created a while ago, simply in order for me to learn a little more about programming with sqlite3 in python.